Refactor: routeHandler를 삭제 => App컴포넌트 생성 후 상태 끌어올림
Stations 컴포넌트에서 관리하던 상태를 Lines 컴포넌트에서도 써야해서 App컴포넌트로 상태를 끌어올려줬음

그 과정에서 IIFE를 썻는데 IIFE도 async함수, IIFE가 반환하는 함수도 async함수여서 await를 한번 해준 뒤 또 await와 함께 함수를 호출해줘야하는 이슈가 있었음

Changes to be committed:
new file:   src/js/app/App.js
modified:   src/js/pages/Stations/Stations.js
modified:   src/js/utils/render.js
deleted:    src/js/utils/routeHandler.js
