+++ src/js/pages/Stations/Stations.js
@@ -1,20 +1,13 @@
-import ajax from 'js/api';
 import { AddStationForm, Heading, StationList } from 'js/components';
 import render from 'js/utils/render';
 
 const Stations = (async () => {
-  let stations = await ajax.getStations();
-  const setStations = newStations => {
-    stations = newStations;
-    render();
-  };
-
   let modalState = { id: null, isModalOpen: false, prevStationName: '' };
   const setModalState = nextModalState => {
     modalState = nextModalState;
     render();
   };
-  return () => {
+  return (stations, setStations) => {
     const $stations = document.createElement('div');
     $stations.className = 'wrapper bg-white p-10';
